添加物料编码转换窗口(电阻电容)。目前已完成： 1.类别代码转换方法；2.耐压，功率代号转换方法；3.精度代号转换方法； 还需要完成： 1.数值转换方法

+++ RuiHaoConvertor/ExcelHelper.cs
@@ -292,11 +292,14 @@ namespace RuiHaoConvertor
         /// </summary>
         public void Dispose()
         {
-            // 杀进程
-            int id = 0;
-            GetWindowThreadProcessId(new IntPtr(_excelApp.Hwnd), out id);
-            System.Diagnostics.Process progress = System.Diagnostics.Process.GetProcessById(id);
-            progress.Kill();
+            if (_excelApp != null)
+            {
+                // 杀进程
+                int id = 0;
+                GetWindowThreadProcessId(new IntPtr(_excelApp.Hwnd), out id);
+                System.Diagnostics.Process progress = System.Diagnostics.Process.GetProcessById(id);
+                progress.Kill();
+            }
         }
         #endregion
